fctorse api call and fetch
