+++ src/components/App/CasesList/CasesList.js
@@ -1,9 +1,11 @@
-import React from 'react'
-import { useSelector } from 'react-redux'
+import React, { useEffect } from 'react'
+import { useSelector, useDispatch } from 'react-redux'
 
 import Container from '@material-ui/core/Container'
 import { makeStyles } from '@material-ui/core/styles'
 
+import { fetchCases } from '../../../services/Cases'
+import { CASES_LIST } from '../../../store/actions'
 import CasesItem from '../CaseItem/CaseItem'
 import titleSvg from '../../../images/maquette/c-case-title.svg'
 
@@ -19,17 +21,26 @@ const useStyles = makeStyles((theme) => ({
 
 const CasesList = () => {
   const classes = useStyles()
+  const dispatch = useDispatch()
+  const home = useSelector((state) => state.home)
   const cases = useSelector((state) => state.home.cases)
-  
+
+  const areLoaded = home.casesLoaded
+
+  useEffect(() => {
+    if (!areLoaded) {
+      fetchCases().then((resp) => (dispatch({ type: CASES_LIST, datas: resp })))
+    }
+  })
+
   return (
     <>
       <Container maxWidth='lg'>
         <center><img src={titleSvg} alt='Cas Cliniques' /></center>
         <div className={classes.root}>
-          { cases && 
-            cases.map((oCase, index) => (
+          { cases && cases.map((oCase, index) => (
               <CasesItem key={index} item={oCase} />
-            )
+              )
             )
           }
         </div>
