+++ src/containers/Cases/Cases.js
@@ -6,10 +6,9 @@ import { makeStyles } from '@material-ui/core/styles'
 
 import Header from '../../components/App/Header/Header'
 import { setup } from '../../services/Auth'
-import { fetchCases } from '../../services/Cases'
-import { CASES_LIST } from '../../store/actions'
+
 import CasesList from '../../components/App/CasesList/CasesList'
-import { useSelector, useDispatch } from 'react-redux'
+
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -20,18 +19,6 @@ const useStyles = makeStyles((theme) => ({
 const Cases = () => {
   const classes = useStyles()
 
-  const dispatch = useDispatch()
-  const home = useSelector((state) => state.home)
-
-  const areLoaded = home.casesLoaded
-
-  useEffect(() => {
-    if (!areLoaded) {
-      const getCases = fetchCases()
-      getCases.then((resp) => (dispatch({ type: CASES_LIST, data: resp })))
-    }
-  })
-
   if (setup() === false) {
     return <Redirect to='/' />
   }
@@ -40,7 +27,7 @@ const Cases = () => {
     <>
       <Header target='home' />
       <Container className={classes.root} />
-      <CasesList />
+      <CasesList/>
     </>
   )
 }
